+++ dmr/internal/django.py
@@ -72,6 +72,7 @@ def convert_multi_value_dict(
     *,
     force_list: frozenset[str],
     cast_null: frozenset[str],
+    split_commas: frozenset[str] | None = None,
 ) -> dict[str, Any]:
     """
     Convert multi value dictionary to a regular one.
@@ -84,7 +85,12 @@ def convert_multi_value_dict(
     Additionally, this function automatically converts the string literal
     ``'null'`` into Python's ``None`` for fields in *cast_null*.
 
-    We use the last value that is sent via query,
+    If *split_commas* is passed, then we also split given field aliases
+    by ``','`` char. Be careful! If data can contain commas as a regular data,
+    it can be corrupted. Use it when you are sure that no commas are possible.
+    For example, with ``list[int]`` data.
+
+    We use the last value that is sent via multivalue dict,
     if there are multiple ones and only one is needed.
     """
     regular_dict: dict[str, Any] = {}
@@ -93,6 +99,15 @@ def convert_multi_value_dict(
             regular_dict[dict_key] = [
                 _replace_null_string(dict_key, list_value, cast_null=cast_null)
                 for list_value in to_parse.getlist(dict_key)
+            ]
+        elif split_commas is not None and dict_key in split_commas:
+            regular_dict[dict_key] = [
+                _replace_null_string(
+                    dict_key,
+                    part,
+                    cast_null=cast_null,
+                )
+                for part in to_parse.get(dict_key, '').split(',')
             ]
         else:
             regular_dict[dict_key] = _replace_null_string(